+++ cythonpowered/dateutil/__init__.py
@@ -1,0 +1,1 @@
+from cythonpowered.dateutil.dateutil import date