+++ utils/definitions/_base.py
@@ -12,16 +12,17 @@ class BaseFunctionDefinition:
 
 class BaseFunctionDefinitionPrinter:
 
-    FUNCTION_PAIRS: List[List[BaseFunctionDefinition]] = []
+    FUNCTION_PAIRS: List[List[List[BaseFunctionDefinition]]] = []
 
     def __init__(self) -> None:
         if len(self.FUNCTION_PAIRS) == 0:
             raise NotImplementedError("No pairs of function definitions provided")
-        for pair in self.FUNCTION_PAIRS:
-            if type(pair) != list:
-                raise TypeError
-            if len(pair) != 2:
-                raise ValueError
+        for module in self.FUNCTION_PAIRS:
+            for pair in module:
+                if type(pair) != list:
+                    raise TypeError
+                if len(pair) != 2:
+                    raise ValueError
         self.print_function_definitions()
 
     def log(self, msg, end: Optional[str] = None) -> None:
@@ -31,20 +32,23 @@ class BaseFunctionDefinitionPrinter:
         table = PrettyTable()
         table.field_names = [
             "#",
-            "[cythonpowered] function",
-            "Replaces [Python] function",
+            "Python function",
+            "Is replaced by",
             "Usage / details",
         ]
         for f in table.field_names:
             table.align[f] = "r" if f == table.field_names[0] else "l"
-
-        for i in range(len(self.FUNCTION_PAIRS)):
-            table.add_row(
-                [
-                    i + 1,
-                    self.FUNCTION_PAIRS[i][1].reference,
-                    self.FUNCTION_PAIRS[i][0].reference,
-                    self.FUNCTION_PAIRS[i][1].usage,
-                ]
-            )
+        i = 0
+        for module in self.FUNCTION_PAIRS:
+            for pair in module:
+                i += 1
+                table.add_row(
+                    [
+                        i,
+                        pair[0].reference,
+                        pair[1].reference,
+                        pair[1].usage,
+                    ]
+                )
+            table.add_divider()
         self.log(table)