+++ utils/definitions/list_functions.py
@@ -1,8 +1,9 @@
 from utils.definitions._base import BaseFunctionDefinitionPrinter
 from utils.definitions._random import RANDOM_DEFINITION_PAIRS
+from utils.definitions._dateutil import DATEUTIL_DEFINITION_PAIRS
 
 
-ALL_PAIRS = RANDOM_DEFINITION_PAIRS
+ALL_PAIRS = [RANDOM_DEFINITION_PAIRS, DATEUTIL_DEFINITION_PAIRS]
 
 
 class AllFunctionDefinitionPrinter(BaseFunctionDefinitionPrinter):